Simple Bayesian Estimates of the Time-Varying Reproduction Number

calibrate_rt() uses the methodology of Cori et al 2013 to estimate the time- varying reproduction number. It calls estimate_R(method = "parametric_si") under the hood. It purposefully does not support aggregating counts over multiple time periods due to poor coverage of the resulting credible intervals, as well as shifting of the results due to lagging of the results. Use the trend argument in prep_linelist() to pre-smooth the data instead.

calibrate_rt(
  .data,
  incid = "trend",
  .t = "collection_date",
  serial_interval_mean = 6,
  serial_interval_sd = 4.17
)

Arguments

.data

A data frame containing the incidence curve and dates

incid

The quoted name of a numeric column containing the incidence curve

.t

The quoted name of a date column corresponding to the observations in incid

serial_interval_mean

The average number of days between infection of a primary case and a secondary case

serial_interval_sd

The standard deviation of the number of days between infection of a primary case and a secondary case

Value

A tibble with columns .t, .pred (the median), .pred_lower(the lower bound of the 95% credible interval), .pred_upper(the upper bound of the 95% credible interval), .mean (the average), and .cv (the coefficient of variation)