Simulate a SIR Model with Vaccination and Variants

sirv() simulates a SIR model with vaccinations, a variant type with increased transmissibility, and a variant type with immunity resistance. For simplicity, vaccine- and naturally-acquired immunity are considered equivalent.

sirv(
  start = Sys.Date(),
  end = Sys.Date() + 60L,
  rt = 1,
  r0 = NULL,
  pop = 937166,
  active_cases = 75,
  total_cases = 90110,
  total_vac = 169150,
  vac_per_day = 3500,
  pct_vac_s = 0.8,
  vac_eff = 0.9,
  vac_eff_ek = 0.6,
  pct_var = 0.1/0.85,
  pct_var_uk = 0.85,
  pct_var_ek = 0.01,
  detect = 1/3
)

Arguments

start
Date or coercible to Date. The start date for simulation.

end
Date or coercible to Date. The end date for simulation.

rt
numeric. The observed reproduction number at the start of simulation; assumed constant other than population and variant effects.

r0
numeric. The baseline reproduction number of the non-variant strains. If this is supplied, rt is ignored.

pop
numeric. The size of the population being modeled.

active_cases
numeric. The observed number of active cases at the start of simulation.

total_cases
numeric. The observed number of total cases at the start of simulation.

total_vac
numeric. The number of vaccinated people at the start of simulation.

vac_per_day
numeric. The number of people vaccinated per day during simulation.

pct_vac_s
numeric between 0 and 1. The fraction of vaccinated people vaccinated who have never had COVID (i.e. are in the susceptible, or "S", compartment).

vac_eff
numeric between 0 and 1. Efficacy of vaccination and natural immunity for non-resistant variants.

pct_var
numeric between 0 and 1. Percent of active cases that are variants of some form.

pct_var_uk
numeric between 0 and 1. Percent of variants that are B.1.1.7 (UK) variant or similarly highly transmissible. Assumed to not overlap with pct_var_ek.

pct_var_ek
numeric between 0 and 1. Percent of variants that possess the E484K mutation (vaccine resistance). Assumed to not overlap with pct_var_uk.

detect
numeric between 0 and 1. The percent of true cases detected; applies to both past and future cases.

vacc_eff_ek
numeric between 0 and 1. Efficacy of vaccination and natural immunity for resistant variants (with the E484K mutation).

Value

A tibble with columns time <date>, S <dbl>, I <dbl>, R <dbl>, pct_uk <dbl>, pct_ek <dbl>